requirements.txt
cssselect2==0.7.0
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.3
django-heroku==0.3.1
django-pdfkit==0.3.1
et-xmlfile==1.1.0
filelock==3.6.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.75.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
googleapis-common-protos==1.58.0
greenlet==1.1.2
gspread==5.2.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
Jinja2==3.0.3
lxml==4.9.1
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.0
oscrypto==1.3.0
Pillow==9.2.0
pyhanko-certvalidator==0.19.5
pyparsing==3.0.7
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.32
sqlparse==0.4.2
svglib==1.4.1
tinycss2==1.1.1
uritemplate==4.1.1
uritools==4.0.0
urllib3==1.26.9
virtualenv==20.13.4
webencodings==0.5.1